tests/wpt/tests/css/css-box/parsing/overflow-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: parsing overflow with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-overflow">
<meta name="assert" content="overflow supports the full grammar '[ visible | hidden | scroll | auto | no-display | no-content ]{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("overflow", "visible");
test_valid_value("overflow", "hidden");
test_valid_value("overflow", "scroll");
test_valid_value("overflow", "auto");

test_valid_value("overflow-x", "visible");
test_valid_value("overflow-x", "hidden");
test_valid_value("overflow-x", "scroll");
test_valid_value("overflow-y", "auto");

// The following are not yet supported by browsers:
// test_valid_value("overflow", "no-display");
// test_valid_value("overflow", "no-content");
// test_valid_value("overflow", "visible hidden");
// test_valid_value("overflow", "auto scroll");
// test_valid_value("overflow", "no-display no-content");
// test_valid_value("overflow-y", "no-display");
// test_valid_value("overflow-y", "no-content");
</script>
</body>
</html>
